requirements.txt
configobj
mysql-connector-python
